Break out GLTexture into its own file
